# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dokumentation der REST-API",
        "title": "Meine API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Returns valid token for a user to use it for the api requests",
                "tags": [
                    "Auth"
                ],
                "summary": "Returns valid token",
                "parameters": [
                    {
                        "description": "Username and password",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AuthInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "description": "Create a user which can be used to create a token",
                "tags": [
                    "Auth"
                ],
                "summary": "Creates a user",
                "parameters": [
                    {
                        "description": "Username and password",
                        "name": "signup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AuthInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/entry": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a list of all entries (fitting to the filters)",
                "tags": [
                    "Entry"
                ],
                "summary": "Get all entries (with filters)",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Filter by done status",
                        "name": "done",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by planned date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EntryDto"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the entry with the given ID. The existing item will be completely overwritten by the provided data. The item will be identified by the ID of the item.",
                "tags": [
                    "Entry"
                ],
                "summary": "Update a entry",
                "parameters": [
                    {
                        "description": "Updated Entry object",
                        "name": "Entry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EntryDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EntryDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new entry",
                "tags": [
                    "Entry"
                ],
                "summary": "Create new entry",
                "parameters": [
                    {
                        "description": "New entry",
                        "name": "entry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateEntryDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.EntryDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/entry/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns exactly one entry with the given ID",
                "tags": [
                    "Entry"
                ],
                "summary": "Get a entry by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the entry",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EntryDto"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes exactly one entry by ID",
                "tags": [
                    "Entry"
                ],
                "summary": "Delete a entry by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the entry",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a list of all entries",
                "tags": [
                    "Entry"
                ],
                "summary": "Get all entries",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EntryDto"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AuthInput": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.CreateEntryDto": {
            "type": "object",
            "properties": {
                "done": {
                    "type": "boolean"
                },
                "plannedDate": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.EntryDto": {
            "type": "object",
            "properties": {
                "done": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "plannedDate": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your token with the 'Bearer ' prefix. Example: 'Bearer eyJhbGciOi...'",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "tags": [
        {
            "description": "All entry-related endpoints",
            "name": "Entry"
        },
        {
            "description": "Login \u0026 registration",
            "name": "Auth"
        }
    ]
}